@@ -1,33 +1,27 @@
 {
   "name": "uchi-widget",
-  "version": "1.2.2",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build && mkdir -p ~/git/cdn/widgets/catalog/v$npm_package_version && cp dist/js/app*.js ~/git/cdn/widgets/catalog/v$npm_package_version/uchi-widget.js && cp dist/css/app*.css ~/git/cdn/widgets/catalog/v$npm_package_version/uchi-widget.css && ln -sfn v$npm_package_version ~/git/cdn/widgets/catalog/v1",
+    "build": "vue-cli-service build && mkdir -p ~/git/cdn/widgets/catalog/v$npm_package_version && cp dist/js/app*.js ~/git/cdn/widgets/catalog/v$npm_package_version/uchi-widget.js && cp dist/css/app*.css ~/git/cdn/widgets/catalog/v$npm_package_version/uchi-widget.css && ln -sfn v$npm_package_version ~/git/cdn/widgets/catalog/v2",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.0.1"
+    "core-js": "^3.8.3",
+    "sass": "^1.57.1",
+    "sass-loader": "^13.2.0",
+    "vue": "^3.2.13",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "autoprefixer": "^9.3.1",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^5.9.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.10.0",
-    "sass-loader": "^7.1.0",
-    "vue-template-compiler": "^2.6.10"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
@@ -35,19 +29,18 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
-    "rules": {
-      "no-console": "off"
-    }
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not dead"
+    "not dead",
+    "not ie 11"
   ]
 }
